feat: split logcat test and add report upload
