+++ integrationtest/requirements.txt
@@ -2,3 +2,4 @@ Appium-Python-Client~=3.1.1
 pytest~=7.4.3
 boto3~=1.34.11
 requests~=2.31.0
+PyYAML~=6.0.1
